refactor(hardhat): make v2 root
