+++ packages/hardhat/test/foundry/KittyFacet.t.sol
@@ -5,10 +5,10 @@ import 'forge-std/Test.sol';
 import {IERC20} from '@openzeppelin/contracts/token/ERC20/IERC20.sol';
 import {FacetTest} from './helpers/FacetTest.sol';
 import {Mainnet} from './helpers/Mainnet.sol';
-import {IDiamondCut} from 'contracts/v2/interfaces/IDiamondCut.sol';
-import {IUniV2Router} from 'contracts/v2/interfaces/IUniV2Router.sol';
-import {KittyFacet} from 'contracts/v2/facets/KittyFacet.sol';
-import {LibKitty} from 'contracts/v2/libraries/LibKitty.sol';
+import {IDiamondCut} from 'contracts/interfaces/IDiamondCut.sol';
+import {IUniV2Router} from 'contracts/interfaces/IUniV2Router.sol';
+import {KittyFacet} from 'contracts/facets/KittyFacet.sol';
+import {LibKitty} from 'contracts/libraries/LibKitty.sol';
 
 contract KittyFacetHarness is KittyFacet {
   function exposed_registerCollectedFee(
