+++ packages/hardhat/test/foundry/SifiDiamond.t.sol
@@ -2,12 +2,12 @@
 pragma solidity ^0.8.21;
 
 import {DiamondHelpers} from './helpers/DiamondHelpers.sol';
-import {SifiDiamond} from 'contracts/v2/SifiDiamond.sol';
-import {DiamondCutFacet} from 'contracts/v2/facets/DiamondCutFacet.sol';
-import {IDiamondCut} from 'contracts/v2/interfaces/IDiamondCut.sol';
-import {DiamondLoupeFacet} from 'contracts/v2/facets/DiamondLoupeFacet.sol';
-import {OwnershipFacet} from 'contracts/v2/facets/OwnershipFacet.sol';
-import {LibDiamond} from 'contracts/v2/libraries/LibDiamond.sol';
+import {SifiDiamond} from 'contracts/SifiDiamond.sol';
+import {DiamondCutFacet} from 'contracts/facets/DiamondCutFacet.sol';
+import {IDiamondCut} from 'contracts/interfaces/IDiamondCut.sol';
+import {DiamondLoupeFacet} from 'contracts/facets/DiamondLoupeFacet.sol';
+import {OwnershipFacet} from 'contracts/facets/OwnershipFacet.sol';
+import {LibDiamond} from 'contracts/libraries/LibDiamond.sol';
 import {Mainnet} from './helpers/Mainnet.sol';
 
 contract SifiDiamondHarness is SifiDiamond {
